Facebook/requirements.txt
selenium
pyyaml
webdriver_manager